+++ lib/lwk.dart
@@ -9,3 +9,4 @@ export './src/generated/api/blockchain.dart';
 export './src/generated/api/transaction.dart';
 export './src/utils/loader.dart';
 export './src/generated/api/error.dart';
+export './src/consolidate.dart';